Added default readme, stickler file, and moved raw html to other file
