+++ README.md
@@ -1,0 +1,51 @@
+# To-Do list
+
+To-Do web application.
+
+## Live demo
+
+[Live website](https://raw.githack.com/Fig77/tic-tac-toejs/init_commit/index.html/)
+
+### Usage
+
+If you want to test and use this project, you need to do:
+* `$ git clone https://github.com/Fig77/tic-tac-toejs/tree/init_commit`
+
+after that you need to run the bundler
+* `$ npm install`
+
+and open your files with live server or similar
+
+
+## Built With
+
+- bootstrap
+- CSS
+- HTML
+- JavaScript
+- Webpack
+
+
+## Authors
+
+👤 **HADDAD SOHAIB**
+
+- Github: [@HADDADSOHAIB](https://github.com/HADDADSOHAIB)
+- Twitter: [@HaddadSohaib](https://twitter.com/HaddadSohaib)
+- Linkedin: [linkedin](https://www.linkedin.com/in/sohaibhaddad/)
+
+👤 **Facundo Iglesias**
+
+- Github: [@Fig77](https://github.com/Fig77)
+- Twitter: [@Fig77](https://twitter.com/Fig_77/)
+- Linkedin: [linkedin](https://www.linkedin.com/in/figlesias/)
+
+## 🤝 Contributing
+
+Contributions, issues and feature requests are welcome!
+
+Feel free to check the [issues page](issues/).
+
+## Show your support
+
+Give a ⭐️ if you like this project!
