+++ package.json
@@ -22,6 +22,7 @@
     "css-loader": "^3.5.3",
     "file-loader": "^6.0.0",
     "jquery": "^3.5.1",
+    "moment": "^2.25.3",
     "popper.js": "^1.16.1",
     "style-loader": "^1.2.1",
     "uid": "^1.0.0",
